refactor iterator and pipeline structure
