+++ vortex/utils/data/augment/modules/nvidia_dali/__init__.py
@@ -1,1 +1,2 @@
 from .modules import *
+from .pipeline import DALIExternalSourcePipeline
