+++ vortex/utils/data/loader/modules/nvidia_dali/__init__.py
@@ -1,0 +1,1 @@
+from .wrapper import DALIIteratorWrapper
