+++ vortex/utils/data/loader/nvidia_dali_loader.py
@@ -1,225 +1,20 @@
-from ..dataset.wrapper import BasicDatasetWrapper, DefaultDatasetWrapper
 from easydict import EasyDict
 from typing import Type, Union, List
 from pathlib import Path
 import random
 import numpy as np
-from collections import OrderedDict
-from nvidia.dali.pipeline import Pipeline
-import nvidia.dali.ops as ops
-import nvidia.dali.types as types
 from nvidia.dali.plugin.pytorch import DALIGenericIterator
 import copy
 import torch
-from ..augment import create_transform
 from ..dataset.wrapper.default_wrapper import check_and_fix_coordinates
 from ....networks.modules.preprocess.normalizer import to_tensor,normalize
 import cv2
 from PIL import Image
 import ray
+from .modules.nvidia_dali import DALIIteratorWrapper
+from ..augment.modules.nvidia_dali import DALIExternalSourcePipeline
+from ..augment import create_transform
 
-
-class DALIIteratorWrapper(object):
-    def __init__(self,
-                 dataset : Type[BasicDatasetWrapper],
-                 batch_size: int,
-                 device_id: int = 0,
-                 num_gpus: int = 1,
-                 shuffle: bool = True):
-        """An extended iterator to bridge dataset and NVIDIA DALI External Input Source Pipeline
-
-        Arguments:
-            dataset {object} -- A base dataset iterator object to return namedtuple data format
-            batch_size {int} -- Batch size of data returned in every iteration
-
-        Keyword Arguments:
-            device_id {int} -- GPU ID to be use (default: {0})
-            num_gpus {int} -- Total GPUs to be used (default: {1})
-            shuffle {bool} -- Shuffle dataset (default: {True})
-
-        Raises:
-            TypeError: if base iterator dataset not returning namedtuple _fields
-            ValueError: if 'image' field is not found on the returned value
-            TypeError: if __len__ method not implemented in base iterator dataset
-        """
-        self.batch_size = batch_size
-        self.dataset = dataset
-
-        # Dataset check
-        if not isinstance(self.dataset,BasicDatasetWrapper) or isinstance(self.dataset,DefaultDatasetWrapper):
-            raise RuntimeError('DALI loader `dataset` args expect BasicDatasetWrapper object, found {}'.format(str(type(dataset))))
-        
-        if not isinstance(self.dataset[0][0],str) and not isinstance(self.dataset[0][0],Path):
-            raise RuntimeError('DALI loader expect `dataset` image data as `str` or `Path` object of the image file, found {}'.format(type(self.dataset[0][0])))
-        
-        # Whole dataset size
-        try:
-            self.dataset_len = len(self.dataset)
-        except:
-            raise RuntimeError(
-                "'dataset' __len__ is not implemented!!")
-        
-        # Pipeline data layout, only include labels with specified `data_format`
-        # `original_labels` is unsliced labels
-        self.data_layout = ['images']+[key for key in self.dataset.data_format.keys() if self.dataset.data_format[key] is not None]+['original_labels']
-
-        # Get proper sharded data based on the device_id and total number of GPUs (world size) (For distributed training in future usage)
-        self.sharded_idx = list(range(self.dataset_len * device_id //
-                                      num_gpus, self.dataset_len * (device_id + 1) // num_gpus))
-        if shuffle:
-            random.shuffle(self.sharded_idx)
-        self.nrof_sharded_data = len(self.sharded_idx)
-
-    def __iter__(self):
-        self.i = 0
-        return self
-
-    def __next__(self):
-        return_data = OrderedDict()
-        for layout in self.data_layout:
-            return_data[layout] = []
-
-        for _ in range(self.batch_size):
-            selected_index = self.sharded_idx[self.i]
-            image,labels = self.dataset[selected_index]
-            sliced_labels = self._slice_labels(labels)
-            for layout in self.data_layout:
-                if layout == 'images':
-                    with open(image, 'rb') as f:
-                        return_data[layout].append(np.frombuffer(f.read(), dtype=np.uint8))
-                elif layout != 'original_labels':
-                    return_data[layout].append(sliced_labels[layout])
-                # Original labels array is also forwarded to add flexibility for labels with unsupported data_format
-                # So it can still be forwarded to the pipeline output
-                else:
-                    if len(labels.shape)==1:
-                        labels=labels[np.newaxis,:]
-                    return_data[layout].append(labels)
-            self.i = (self.i + 1) % self.nrof_sharded_data
-        return tuple(return_data[layout] for layout in self.data_layout)
-
-    @property
-    def size(self,):
-        return self.nrof_sharded_data
-
-    def _slice_labels(self,labels):
-        sliced_labels = {}
-
-        for label_name in self.dataset.data_format:
-            if self.dataset.data_format[label_name] is not None:
-                sliced_labels[label_name]= np.take(labels, 
-                                                axis=self.dataset.data_format[label_name].axis, 
-                                                indices=self.dataset.data_format[label_name].indices)
-                
-                # Transform flattened landmarks coords into 2 dim array
-                if label_name=='landmarks':
-                    sliced_labels[label_name]=sliced_labels[label_name].reshape((int(sliced_labels[label_name].size/2), 2))
-        
-        ## Handle if labels contain only 'class_labels' and the data format is None
-        return sliced_labels
-
-    next = __next__
-
-class DALIExternalSourcePipeline(Pipeline):
-    def __init__(self, dataset_iterator, 
-                       batch_size, 
-                       num_threads, 
-                       device_id, 
-                       dali_augments = None,
-                       normalize = True,
-                       seed = 12345):
-        super().__init__(
-                        batch_size,
-                        num_threads,
-                        device_id,
-                        seed=seed)
-        self.original_data_layout = copy.copy(dataset_iterator.data_layout)
-        self.pipeline_output_data_layout = copy.copy(dataset_iterator.data_layout)
-        self.source = ops.ExternalSource(source = dataset_iterator, 
-                                         num_outputs = len(dataset_iterator.data_layout))
-        self.image_decode = ops.ImageDecoder(device = "mixed", output_type = types.BGR)
-        self.label_cast = ops.Cast(device="cpu",
-                             dtype=types.FLOAT)
-        self.labels_pad_value = -99
-
-        # Nvidia DALI augmentations from experiment file
-        self.dali_augments = dali_augments
-
-        # Modify pipeline output data layout to prepare placeholder for required information
-        # that need to be passed from DALI pipeline to DALI loader
-        
-        ## Information placeholder for flip augmentation flag used for asymmetric information
-        flag_count = 0
-        if self.dali_augments:
-            for augment in self.dali_augments.compose:
-
-                # Somehow, VerticalFlip in DALI didn't break the sequence for asymm landmarks
-                if type(augment).__name__ == 'HorizontalFlip' or type(augment).__name__ == 'VerticalFlip':
-                    self.pipeline_output_data_layout.append('flip_flag_'+str(flag_count))
-                    flag_count+=1
-
-        ## Information placeholder for standard augment used for coordinates fixing
-        self.pipeline_output_data_layout.append('pre_padded_image_shape')
-        # Standard Augments Resize and Pad to Square
-        preprocess_args = dataset_iterator.dataset.preprocess_args
-        if 'mean' not in preprocess_args.input_normalization:
-            preprocess_args.input_normalization.mean=[0.,0.,0.]
-        if 'std' not in preprocess_args.input_normalization:
-            preprocess_args.input_normalization.mean=[1.,1.,1.]
-        if 'scaler' not in preprocess_args.input_normalization:
-            preprocess_args.input_normalization.scaler=255
-
-        standard_augment = EasyDict()
-        standard_augment.data_format = dataset_iterator.dataset.data_format
-        standard_augment.transforms = [
-            {'transform': 'StandardAugment','args':{'scaler' : preprocess_args.input_normalization.scaler,
-                                                    'mean' : preprocess_args.input_normalization.mean,
-                                                    'std' : preprocess_args.input_normalization.std,
-                                                    'input_size' : preprocess_args.input_size,
-                                                    'image_pad_value' : 0,
-                                                    'labels_pad_value' : self.labels_pad_value,
-                                                    'normalize' : normalize
-                                                    }
-            }
-        ]
-        self.standard_augments = create_transform(
-            'nvidia_dali', **standard_augment)
-
-    def define_graph(self):
-        pipeline_input = self.source()
-
-        # Prepare pipeline input data
-        data = EasyDict()
-        graph_data=dict(zip(self.original_data_layout, pipeline_input))
-        data.images = self.image_decode(graph_data['images'])
-        data.labels = EasyDict()
-        for layout in self.original_data_layout:
-            if layout!='images':
-                data.labels[layout]=self.label_cast(graph_data[layout])
-        data.additional_info=EasyDict()
-        data.data_layout=self.original_data_layout
-        # Custom Augmentation Start
-
-        if self.dali_augments:
-            data = self.dali_augments(**data)
-
-        # Custom Augmentation End
-        data = self.standard_augments(**data)
-
-        # Prepare pipeline output tuple
-        pipeline_output = []
-
-        # Iterate new layout
-        new_layout = data.data_layout
-
-        for component in new_layout:
-            if component == 'images':
-                pipeline_output.append(data[component])
-            elif component in data.labels:
-                pipeline_output.append(data.labels[component])
-            elif component in data.additional_info:
-                pipeline_output.append(data.additional_info[component])
-        return pipeline_output
 
 class DALIDataloader():
     def __init__(self,
